agent-proxy: Squelch compile warnings
agent-proxy.c: In function ‘writeScriptClients’:
agent-proxy.c:806:6: warning: variable ‘ret’ set but not used [-Wunused-but-set-variable]
agent-proxy.c: In function ‘main’:
agent-proxy.c:1460:6: warning: variable ‘latency’ set but not used [-Wunused-but-set-variable]
agent-proxy.c:1459:6: warning: variable ‘baud’ set but not used [-Wunused-but-set-variable]

The ret warning just needs to have the variable deleted.   For the latency
and baud, they are for a seldom used feature that is #if def'ed and
these variables should be behind the #if USE_LATENCY option.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>